refactor(e2e): promote TxSender, governance join, layout asserts and upgrade_lib to shared helpers
Extractions the storage_v2_fresh_sync case (TC9) reuses, per the design's
reuse map:
- helpers/tx_sender.py: third TxSender occurrence; adds pause()/resume()
  for windows where the chain is expected to freeze (2-validator swap
  windows, the L3 necessity probe) and makes the fallback target optional.
- helpers/governance.py (H4): the fuzzy_cluster one-time permissionless-
  join proposal lifecycle, parameterized by faucet key/address.
- storage_case_lib.assert_sf_layout / assert_upgraded_legacy_layout: the
  offline layout criteria shared by TC4 and TC9 (probe callables
  injectable for unit tests).
- helpers/upgrade_lib.py: promoted from the storage_v2_upgrade case dir
  (compat shim left behind) so the rolling-upgrade recipe is importable
  from other cases.

storage_v2_upgrade and fuzzy_cluster refactored onto the shared code;
no behavior change intended.
